splited controller on two classes MeasurementsController and InstallationsController
