--- .idea/libraries/Maven__com_google_code_gson_gson_2_8_6.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.google.code.gson:gson:2.8.6">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.8.6/gson-2.8.6.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.8.6/gson-2.8.6-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.8.6/gson-2.8.6-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
