--- .idea/libraries/Maven__com_squareup_okhttp3_okhttp_3_14_7.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.squareup.okhttp3:okhttp:3.14.7">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okhttp3/okhttp/3.14.7/okhttp-3.14.7.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okhttp3/okhttp/3.14.7/okhttp-3.14.7-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okhttp3/okhttp/3.14.7/okhttp-3.14.7-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
