--- .idea/libraries/Maven__com_squareup_okio_okio_1_17_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.squareup.okio:okio:1.17.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okio/okio/1.17.2/okio-1.17.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okio/okio/1.17.2/okio-1.17.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/okio/okio/1.17.2/okio-1.17.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
