--- .idea/libraries/Maven__com_squareup_retrofit2_converter_gson_2_3_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.squareup.retrofit2:converter-gson:2.3.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/converter-gson/2.3.0/converter-gson-2.3.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/converter-gson/2.3.0/converter-gson-2.3.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/converter-gson/2.3.0/converter-gson-2.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
