--- .idea/libraries/Maven__com_squareup_retrofit2_retrofit_2_5_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.squareup.retrofit2:retrofit:2.5.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/retrofit/2.5.0/retrofit-2.5.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/retrofit/2.5.0/retrofit-2.5.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/squareup/retrofit2/retrofit/2.5.0/retrofit-2.5.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
