+++ src/main/java/com/agnieszka/piotrowska/weatherCheckApp/model/MeasurementHistory.java
@@ -2,7 +2,6 @@ package com.agnieszka.piotrowska.weatherCheckApp.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
-
 import java.util.List;
 
 @JsonIgnoreProperties
