+++ src/main/java/com/agnieszka/piotrowska/weatherCheckApp/model/SingleMeasurement.java
@@ -2,7 +2,6 @@ package com.agnieszka.piotrowska.weatherCheckApp.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
-
 import java.util.List;
 
 @JsonIgnoreProperties
@@ -11,7 +10,7 @@ public class SingleMeasurement {
 
     private String fromDateTime;
     private String tillDateTime;
-    private List<Value> values;
+    private List <Value> values;
     private List <Index> indexes;
     private List <Standards> standards;
